+++ docs/metro/js/metro.js
@@ -106,8 +106,8 @@ var isTouch = (('ontouchstart' in window) || (navigator.MaxTouchPoints > 0) || (
 
 var Metro = {
 
-    version: "4.2.39-dev 13/03/2019 21:05",
-    versionFull: "4.2.39-dev 13/03/2019 21:05",
+    version: "4.2.39-dev 15/03/2019 17:04",
+    versionFull: "4.2.39-dev 15/03/2019 17:04",
     isTouchable: isTouch,
     fullScreenEnabled: document.fullscreenEnabled,
     sheet: null,
@@ -19309,6 +19309,7 @@ var Table = {
         var search_block, search_input, rows_block, rows_select;
 
         search_block = Utils.isValue(this.wrapperSearch) ? this.wrapperSearch : $("<div>").addClass("table-search-block").addClass(o.clsSearch).appendTo(top_block);
+        search_block.addClass(o.clsSearch);
 
         search_input = $("<input>").attr("type", "text").appendTo(search_block);
         search_input.input({
@@ -19319,7 +19320,8 @@ var Table = {
             search_block.hide();
         }
 
-        rows_block = Utils.isValue(this.wrapperRows) ? this.wrapperRows : $("<div>").addClass("table-rows-block").addClass(o.clsRowsCount).appendTo(top_block);
+        rows_block = Utils.isValue(this.wrapperRows) ? this.wrapperRows : $("<div>").addClass("table-rows-block").appendTo(top_block);
+        rows_block.addClass(o.clsRowsCount);
 
         rows_select = $("<select>").appendTo(rows_block);
         $.each(Utils.strToArray(o.rowsSteps), function () {
@@ -19356,12 +19358,14 @@ var Table = {
         var bottom_block = $("<div>").addClass("table-bottom").addClass(o.clsTableBottom).insertAfter(element.parent());
         var info, pagination;
 
-        info = Utils.isValue(this.wrapperInfo) ? this.wrapperInfo : $("<div>").addClass("table-info").addClass(o.clsTableInfo).appendTo(bottom_block);
+        info = Utils.isValue(this.wrapperInfo) ? this.wrapperInfo : $("<div>").addClass("table-info").appendTo(bottom_block);
+        info.addClass(o.clsTableInfo);
         if (o.showTableInfo !== true) {
             info.hide();
         }
 
-        pagination = Utils.isValue(this.wrapperPagination) ? this.wrapperPagination : $("<div>").addClass("table-pagination").addClass(o.clsTablePagination).appendTo(bottom_block);
+        pagination = Utils.isValue(this.wrapperPagination) ? this.wrapperPagination : $("<div>").addClass("table-pagination").appendTo(bottom_block);
+        pagination.addClass(o.clsTablePagination);
         if (o.showPagination !== true) {
             pagination.hide();
         }
@@ -22331,7 +22335,7 @@ var Touch = {
     },
 
     options: {
-        fingers: 1,
+        fingers: 2,
         threshold: 75,
         cancelThreshold: null,
         pinchThreshold: 20,