+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 + [x] Table: fix assign classes from attribute `data-cls-head-cell`
 + [x] Breadcrumbs: add `.breadcrumb-item` class
 + [x] Draggable: fix recreate element content when drag started
++ [x] Table: fix implements custom class to custom wrappers (search, pagination, ...), issue #1335
 
 ### 4.2.38
 + [x] ListView: add class `.vertical-layout` for `icons-*` view mode